backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
motor==3.4.0
pymongo==4.7.3
pydantic==2.8.2
pydantic-settings==2.4.0
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pytest==8.3.2
httpx==0.27.0
black==24.8.0
flake8==7.1.1
openpyxl==3.1.5

